+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- Added ability to get metrics from cloudfront (@yuri-zubov sponsored by Actility, https://www.actility.com)
 
 ## [11.1.0] - 2018-11-22
 ### Added